+++ adm_program/modules/announcements/announcements.php
@@ -24,9 +24,9 @@ require_once('../../system/common.php');
 unset($_SESSION['announcements_request']);
 
 // Initialize and check the parameters
-$getStart    = admFuncVariableIsValid($_GET, 'start',     'numeric');
+$getStart    = admFuncVariableIsValid($_GET, 'start',     'int');
 $getHeadline = admFuncVariableIsValid($_GET, 'headline',  'string', array('defaultValue' => $gL10n->get('ANN_ANNOUNCEMENTS')));
-$getId       = admFuncVariableIsValid($_GET, 'id',        'numeric');
+$getId       = admFuncVariableIsValid($_GET, 'id',        'int');
 $getDateFrom = admFuncVariableIsValid($_GET, 'date_from', 'date');
 $getDateTo   = admFuncVariableIsValid($_GET, 'date_to',   'date');
