+++ adm_program/modules/categories/categories_new.php
@@ -27,7 +27,7 @@ require_once('../../system/common.php');
 require_once('../../system/login_valid.php');
 
 // Initialize and check the parameters
-$getCatId = admFuncVariableIsValid($_GET, 'cat_id', 'numeric');
+$getCatId = admFuncVariableIsValid($_GET, 'cat_id', 'int');
 $getType  = admFuncVariableIsValid($_GET, 'type',   'string', array('requireValue' => true, 'validValues' => array('ROL', 'LNK', 'USF', 'DAT', 'INF', 'AWA')));
 $getTitle = admFuncVariableIsValid($_GET, 'title',  'string');
