+++ adm_program/modules/dates/ical_dates.php
@@ -31,7 +31,7 @@ unset($_SESSION['dates_request']);
 // Initialize and check the parameters
 $getMode     = admFuncVariableIsValid($_GET, 'mode',     'string', array('defaultValue' => 'actual', 'validValues' => array('actual', 'old', 'all')));
 $getHeadline = admFuncVariableIsValid($_GET, 'headline', 'string', array('defaultValue' => $gL10n->get('DAT_DATES')));
-$getCatId    = admFuncVariableIsValid($_GET, 'cat_id',   'numeric');
+$getCatId    = admFuncVariableIsValid($_GET, 'cat_id',   'int');
 
 // prüfen ob das Modul überhaupt aktiviert ist
 if($gPreferences['enable_dates_module'] == 0)
