+++ adm_program/modules/guestbook/guestbook_comment_new.php
@@ -18,8 +18,8 @@
 require_once('../../system/common.php');
 
 // Initialize and check the parameters
-$getGboId    = admFuncVariableIsValid($_GET, 'id',       'numeric');
-$getGbcId    = admFuncVariableIsValid($_GET, 'cid',      'numeric');
+$getGboId    = admFuncVariableIsValid($_GET, 'id',       'int');
+$getGbcId    = admFuncVariableIsValid($_GET, 'cid',      'int');
 $getHeadline = admFuncVariableIsValid($_GET, 'headline', 'string', array('defaultValue' => $gL10n->get('GBO_GUESTBOOK')));
 
 // pruefen ob das Modul ueberhaupt aktiviert ist
