+++ adm_program/modules/members/profile_field_history.php
@@ -24,7 +24,7 @@ $filterDateFrom = new DateTimeExtended(DATE_NOW, 'Y-m-d');
 $filterDateFrom->modify('-'.$gPreferences['members_days_field_history'].' day');
 
 // Initialize and check the parameters
-$getUserId   = admFuncVariableIsValid($_GET, 'usr_id',           'numeric');
+$getUserId   = admFuncVariableIsValid($_GET, 'usr_id',           'int');
 $getDateFrom = admFuncVariableIsValid($_GET, 'filter_date_from', 'date', array('defaultValue' => $filterDateFrom->format($gPreferences['system_date'])));
 $getDateTo   = admFuncVariableIsValid($_GET, 'filter_date_to',   'date', array('defaultValue' => DATE_NOW));
