+++ adm_program/modules/photos/photo_download.php
@@ -23,8 +23,8 @@ require_once('../../system/common.php');
 require_once('../../system/login_valid.php');
 
 // Initialize and check the parameters
-$getPhotoId = admFuncVariableIsValid($_GET, 'pho_id',   'numeric');
-$getPhotoNr = admFuncVariableIsValid($_GET, 'photo_nr', 'numeric');
+$getPhotoId = admFuncVariableIsValid($_GET, 'pho_id',   'int');
+$getPhotoNr = admFuncVariableIsValid($_GET, 'photo_nr', 'int');
 
 // tempfolder
 // change this value if your provider requires the usage of special directories (e.g. HostEurope)
