+++ adm_program/modules/profile/profile_function.php
@@ -26,10 +26,10 @@ require_once('../../system/login_valid.php');
 require_once('roles_functions.php');
 
 // Initialize and check the parameters
-$getUserId   = admFuncVariableIsValid($_GET, 'user_id', 'numeric');
-$getRoleId   = admFuncVariableIsValid($_GET, 'rol_id',  'numeric');
-$getMemberId = admFuncVariableIsValid($_GET, 'mem_id',  'numeric');
-$getMode     = admFuncVariableIsValid($_GET, 'mode',    'numeric');
+$getUserId   = admFuncVariableIsValid($_GET, 'user_id', 'int');
+$getRoleId   = admFuncVariableIsValid($_GET, 'rol_id',  'int');
+$getMemberId = admFuncVariableIsValid($_GET, 'mem_id',  'int');
+$getMode     = admFuncVariableIsValid($_GET, 'mode',    'int');
 
 // in ajax mode only return simple text on error
 if($getMode === 7)
