+++ adm_program/modules/roles/roles_function.php
@@ -29,8 +29,8 @@ require_once('../../system/common.php');
 require_once('../../system/login_valid.php');
 
 // Initialize and check the parameters
-$getRoleId = admFuncVariableIsValid($_GET, 'rol_id', 'numeric');
-$getMode   = admFuncVariableIsValid($_GET, 'mode',   'numeric', array('requireValue' => true));
+$getRoleId = admFuncVariableIsValid($_GET, 'rol_id', 'int');
+$getMode   = admFuncVariableIsValid($_GET, 'mode',   'int', array('requireValue' => true));
 
 // only members who are allowed to create and edit roles should have access to
 // most of these functions
